Simplify notifcation LED handling
The LED will now switch state after each event. When an odd number of
buttons is in the pressed state, the LED will be on. When an even number
of buttons is in the pressed state, the LED will be off.

This simplification also removed the need for the `asyncio` module.